Split pages mitations into multiple files
