+++ saleor/graphql/page/mutations/__init__.py
@@ -1,0 +1,15 @@
+from .page_create import PageCreate
+from .page_delete import PageDelete
+from .page_type_create import PageTypeCreate
+from .page_type_delete import PageTypeDelete
+from .page_type_update import PageTypeUpdate
+from .page_update import PageUpdate
+
+__all__ = [
+    "PageCreate",
+    "PageDelete",
+    "PageUpdate",
+    "PageTypeCreate",
+    "PageTypeUpdate",
+    "PageTypeDelete",
+]
